fix: move styling files out of pages directory
