+++ src/pages/index.tsx
@@ -5,7 +5,7 @@ import useTranslation from '@/hooks/useTranslation'
 
 import { Footer, Activities, AboutUs, Sponsors, Banner } from '@/components'
 
-import * as S from './home.styles'
+import * as S from '../styles/home.styles'
 
 export default function Home() {
   const t = useTranslation()
